% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GFMTW.R
\name{DGP1}
\alias{DGP1}
\title{A function that generate the poisson data matrices with no intercepts}
\usage{
DGP1(N, P, R2)
}
\arguments{
\item{N}{the number of rows}

\item{P}{the number of columns}

\item{R2}{the number of factors for alphai2}
}
\value{
a data matrix named X
}
\description{
A function that generate the poisson data matrices which don't have the intercept term. Details are described in the simulation section.
}
\examples{
\dontrun{
X<-DGP1(200,300,1)
X
}
}
