% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GFMTW.R
\name{DGP2}
\alias{DGP2}
\title{A function that generate the poisson data matrices with intercepts}
\usage{
DGP2(N, P, R2)
}
\arguments{
\item{N}{the number of rows}

\item{P}{the number of columns}

\item{R2}{the number of factors}
}
\value{
a data matrix named X
}
\description{
A function that generate the poisson data matrices which have the intercept term. Details are described in the simulation section.
}
\examples{
\dontrun{
X<-DGP2(200,300,3)
X
}
}
