% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GFMTW.R
\name{DGP_miss}
\alias{DGP_miss}
\title{A function that constructs matrices with missing values}
\usage{
DGP_miss(Nm, Pm, X)
}
\arguments{
\item{Nm}{the number of rows with missing values}

\item{Pm}{the number of columns with missing values}

\item{X}{a N times P data matrix}
}
\value{
a matrix with missing values
}
\description{
A function that constructs matrices with missing values, the result will contain Nm rows with missing values and Pm columns with missing values.
}
\examples{
\dontrun{
X<-matrix(rnorm(20*30),nrow=20,ncol=30)
X_miss<-DGP_miss(10,10,X)
X_miss
}
}
