% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GFMTW.R
\name{GFM_TW}
\alias{GFM_TW}
\title{A function that perform GFM-TW algorithm}
\usage{
GFM_TW(X_miss, num_factor, type = "poission", detail = "no intercept")
}
\arguments{
\item{X_miss}{the data matrix with missing values}

\item{num_factor}{the number of factors for alphai2}

\item{type}{the type of data}

\item{detail}{the details about the data type}
}
\value{
a complete data matrix
}
\description{
A function that perform GFM-TW algorithm for the poisson data. Details see the section simulation.
}
\examples{
\dontrun{
X<-DGP1(200,300,1)
X_miss<-DGP_miss(80,80,X)
X_complete<-GFM_TW(X_miss,3,'poisson','no intercept')
}
}
