% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GFMTW.R
\name{HW0data1}
\alias{HW0data1}
\title{A dataset used for HW0 in vignettes.}
\description{
This dataset is used in HW0 for plots.
}
\examples{
\dontrun{
data(HW0data1)
attach(HW0data1)
xt <- ts(t(HW0data1), start = c(2000,1), frequency = 4) 
ts.plot(xt, main="The average income in USA", ylab="income")
}
}
