% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GFMTW.R
\name{HW0data2}
\alias{HW0data2}
\title{A dataset used for HW0 in vignettes.}
\description{
This dataset is used in HW0 for plots.
}
\examples{
\dontrun{
data(HW0data2)
attach(HW0data2)
data_time <- data.frame(day = ymd(HW0data2[,1]), lowtemp =HW0data2[,3])
ggplot(data_time, aes(x=day, y=lowtemp))+
geom_line(color="blue") + xlab("")+
scale_x_date(date_breaks = "2 years")+
labs(title="The lowest temperature in Hefei")+
theme(axis.text.x=element_text(angle=30,hjust=0.8,vjust=0.8))
}
}
