% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RMSEC}
\alias{RMSEC}
\title{A Rcpp function to calculate RMSE}
\usage{
RMSEC(Xhat, X, missnum)
}
\arguments{
\item{Xhat}{the estimated matrix}

\item{X}{the origin data matrix}

\item{missnum}{the number of missing values in X_miss}
}
\value{
the value of RMSE. Details see Section Simulation.
}
\description{
A Rcpp function to calculate RMSE
}
\examples{
\dontrun{
X1<-matrix(rnorm(20*30),nrow=20,ncol=30)
X1hat<-matrix(rnorm(20*30),nrow=20,ncol=30)
missnum1<-10*10
rmse<-RMSEC(X1hat,X1,missnum1)
}
}
